Draft for a reactive way to calculate trades for tickers received via websockets Refactored SpreadService, TradingService, ExchangeService, ExchangeService. Added class TradingScheduler
