+++ src/main/java/com/r307/arbitrader/service/SpreadService.java
@@ -1,17 +1,22 @@
 package com.r307.arbitrader.service;
 
 import com.r307.arbitrader.service.model.Spread;
+import com.r307.arbitrader.service.model.TradeCombination;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.currency.CurrencyPair;
+import org.knowm.xchange.dto.marketdata.Ticker;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Collectors;
+
+import static com.r307.arbitrader.DecimalConstants.BTC_SCALE;
 
 @Component
 public class SpreadService {
@@ -21,6 +26,11 @@ public class SpreadService {
     private final Map<String, BigDecimal> maxSpreadIn = new HashMap<>();
     private final Map<String, BigDecimal> minSpreadOut = new HashMap<>();
     private final Map<String, BigDecimal> maxSpreadOut = new HashMap<>();
+    private final TickerService tickerService;
+
+    public SpreadService(TickerService tickerService) {
+        this.tickerService = tickerService;
+    }
 
     public void publish(Spread spread) {
         String spreadKey = spreadKey(spread.getLongExchange(), spread.getShortExchange(), spread.getCurrencyPair());
@@ -37,6 +47,42 @@ public class SpreadService {
         LOGGER.info("Maximum spreadIns:\n{}", buildSummary(maxSpreadIn));
         LOGGER.info("Minimum spreadOuts:\n{}", buildSummary(minSpreadOut));
         LOGGER.info("Maximum spreadOuts:\n{}", buildSummary(maxSpreadOut));
+    }
+
+    public Spread computeSpread(TradeCombination tradeCombination) {
+        Exchange longExchange = tradeCombination.getLongExchange();
+        Exchange shortExchange = tradeCombination.getShortExchange();
+        CurrencyPair currencyPair = tradeCombination.getCurrencyPair();
+
+        Ticker longTicker = tickerService.getTicker(longExchange, currencyPair);
+        Ticker shortTicker = tickerService.getTicker(shortExchange, currencyPair);
+
+        if (tickerService.isInvalidTicker(longTicker) || tickerService.isInvalidTicker(shortTicker)) {
+            return null;
+        }
+
+        BigDecimal spreadIn = computeSpread(longTicker.getAsk(), shortTicker.getBid());
+        BigDecimal spreadOut = computeSpread(longTicker.getBid(), shortTicker.getAsk());
+
+        Spread spread = new Spread(
+            currencyPair,
+            longExchange,
+            shortExchange,
+            longTicker,
+            shortTicker,
+            spreadIn,
+            spreadOut);
+
+        publish(spread);
+
+        return spread;
+    }
+
+    public BigDecimal computeSpread(BigDecimal longPrice, BigDecimal shortPrice) {
+        BigDecimal scaledLongPrice = longPrice.setScale(BTC_SCALE, RoundingMode.HALF_EVEN);
+        BigDecimal scaledShortPrice = shortPrice.setScale(BTC_SCALE, RoundingMode.HALF_EVEN);
+
+        return (scaledShortPrice.subtract(scaledLongPrice)).divide(scaledLongPrice, RoundingMode.HALF_EVEN);
     }
 
     private String buildSummary(Map<String, BigDecimal> spreadMap) {
