+++ src/main/java/com/r307/arbitrader/service/TradingService.java
@@ -7,34 +7,21 @@ import com.r307.arbitrader.exception.OrderNotFoundException;
 import com.r307.arbitrader.service.model.ActivePosition;
 import com.r307.arbitrader.service.model.ArbitrageLog;
 import com.r307.arbitrader.service.model.Spread;
-import com.r307.arbitrader.service.model.TradeCombination;
-import com.r307.arbitrader.service.ticker.TickerStrategy;
-import info.bitrich.xchangestream.core.StreamingExchangeFactory;
 import org.apache.commons.io.FileUtils;
 import org.knowm.xchange.Exchange;
-import org.knowm.xchange.ExchangeFactory;
-import org.knowm.xchange.ExchangeSpecification;
 import org.knowm.xchange.currency.Currency;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.Order;
-import org.knowm.xchange.dto.account.Fee;
-import org.knowm.xchange.dto.account.Wallet;
 import org.knowm.xchange.dto.marketdata.OrderBook;
-import org.knowm.xchange.dto.marketdata.Ticker;
 import org.knowm.xchange.dto.meta.CurrencyPairMetaData;
 import org.knowm.xchange.dto.trade.LimitOrder;
 import org.knowm.xchange.dto.trade.OpenOrders;
 import org.knowm.xchange.exceptions.ExchangeException;
 import org.knowm.xchange.exceptions.NotAvailableFromExchangeException;
-import org.knowm.xchange.exceptions.NotYetImplementedForExchangeException;
-import org.knowm.xchange.service.account.AccountService;
-import org.knowm.xchange.service.marketdata.params.CurrencyPairsParam;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.PostConstruct;
 import java.io.File;
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -42,22 +29,16 @@ import java.math.RoundingMode;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.time.OffsetDateTime;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static com.r307.arbitrader.DecimalConstants.BTC_SCALE;
-import static com.r307.arbitrader.DecimalConstants.USD_SCALE;
 
 @Component
 public class TradingService {
-    public static final String METADATA_KEY = "arbitrader-metadata";
-    public static final String TICKER_STRATEGY_KEY = "tickerStrategy";
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(TradingService.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(TradingScheduler.class);
     private static final String STATE_FILE = ".arbitrader/arbitrader-state.json";
     protected static final String TRADE_HISTORY_FILE = ".arbitrader/arbitrader-arbitrage-history.csv";
 
@@ -69,277 +50,73 @@ public class TradingService {
 
     private final ObjectMapper objectMapper;
     private final TradingConfiguration tradingConfiguration;
-    private final ExchangeFeeCache feeCache;
     private final ConditionService conditionService;
     private final ExchangeService exchangeService;
-    private final ErrorCollectorService errorCollectorService;
     private final SpreadService spreadService;
     private final TickerService tickerService;
     private final NotificationService notificationService;
-    private final Map<String, TickerStrategy> tickerStrategies;
-    private final List<Exchange> exchanges = new ArrayList<>();
-    private boolean bailOut = false;
     private boolean timeoutExitWarning = false;
     private ActivePosition activePosition = null;
+    private boolean bailOut = false;
 
     public TradingService(
         ObjectMapper objectMapper,
         TradingConfiguration tradingConfiguration,
-        ExchangeFeeCache feeCache,
         ConditionService conditionService,
         ExchangeService exchangeService,
-        ErrorCollectorService errorCollectorService,
         SpreadService spreadService,
         TickerService tickerService,
-        NotificationService notificationService,
-        Map<String, TickerStrategy> tickerStrategies) {
+        NotificationService notificationService) {
 
         this.objectMapper = objectMapper;
         this.tradingConfiguration = tradingConfiguration;
-        this.feeCache = feeCache;
         this.conditionService = conditionService;
         this.exchangeService = exchangeService;
-        this.errorCollectorService = errorCollectorService;
         this.spreadService = spreadService;
         this.tickerService = tickerService;
-        this.tickerStrategies = tickerStrategies;
         this.notificationService = notificationService;
     }
 
-    @PostConstruct
-    public void connectExchanges() {
-        tradingConfiguration.getExchanges().forEach(exchangeMetadata -> {
-            ExchangeSpecification specification = new ExchangeSpecification(exchangeMetadata.getExchangeClass());
+    public void startTradingProcess() {
+        tickerService.getTradeCombinations()
+            .forEach(tradeCombination -> {
+                Spread spread = spreadService.computeSpread(tradeCombination);
 
-            specification.setUserName(exchangeMetadata.getUserName());
-            specification.setApiKey(exchangeMetadata.getApiKey());
-            specification.setSecretKey(exchangeMetadata.getSecretKey());
-
-            if (exchangeMetadata.getSslUri() != null) {
-                specification.setSslUri(exchangeMetadata.getSslUri());
-            }
-
-            if (exchangeMetadata.getHost() != null) {
-                specification.setHost(exchangeMetadata.getHost());
-            }
-
-            if (exchangeMetadata.getPort() != null) {
-                specification.setPort(exchangeMetadata.getPort());
-            }
-
-            if (!exchangeMetadata.getCustom().isEmpty()) {
-                exchangeMetadata.getCustom().forEach((key, value) -> {
-                    if ("true".equals(value) || "false".equals(value)) {
-                        specification.setExchangeSpecificParametersItem(key, Boolean.valueOf(value));
-                    } else {
-                        specification.setExchangeSpecificParametersItem(key, value);
-                    }
-                });
-            }
-
-            specification.setExchangeSpecificParametersItem(METADATA_KEY, exchangeMetadata);
-
-            if (specification.getExchangeClassName().contains("Streaming")) {
-                exchanges.add(StreamingExchangeFactory.INSTANCE.createExchange(specification));
-            } else {
-                exchanges.add(ExchangeFactory.INSTANCE.createExchange(specification));
-            }
-        });
-
-        exchanges.forEach(exchange -> {
-            try {
-                LOGGER.debug("{} SSL URI: {}",
-                        exchange.getExchangeSpecification().getExchangeName(),
-                        exchange.getExchangeSpecification().getSslUri());
-                LOGGER.debug("{} SSL host: {}",
-                        exchange.getExchangeSpecification().getExchangeName(),
-                        exchange.getExchangeSpecification().getHost());
-                LOGGER.debug("{} SSL port: {}",
-                        exchange.getExchangeSpecification().getExchangeName(),
-                        exchange.getExchangeSpecification().getPort());
-                LOGGER.debug("{} home currency: {}",
-                        exchange.getExchangeSpecification().getExchangeName(),
-                        exchangeService.getExchangeHomeCurrency(exchange));
-                LOGGER.info("{} balance: {}{}",
-                        exchange.getExchangeSpecification().getExchangeName(),
-                    exchangeService.getExchangeHomeCurrency(exchange).getSymbol(),
-                        getAccountBalance(exchange));
-            } catch (IOException e) {
-                LOGGER.error("Unable to fetch account balance: ", e);
-            }
-
-            if (exchange.getExchangeSpecification().getExchangeClassName().contains("Streaming")) {
-                exchange.getExchangeSpecification().setExchangeSpecificParametersItem(TICKER_STRATEGY_KEY, tickerStrategies.get("streamingTickerStrategy"));
-            } else {
-                try {
-                    CurrencyPairsParam param = () -> exchangeService.getExchangeMetadata(exchange).getTradingPairs().subList(0, 1);
-                    exchange.getMarketDataService().getTickers(param);
-
-                    exchange.getExchangeSpecification().setExchangeSpecificParametersItem(TICKER_STRATEGY_KEY, tickerStrategies.get("singleCallTickerStrategy"));
-                } catch (NotYetImplementedForExchangeException e) {
-                    LOGGER.warn("{} does not support fetching multiple tickers at a time and will fetch tickers " +
-                            "individually instead. This may result in API rate limiting.",
-                        exchange.getExchangeSpecification().getExchangeName());
-
-                    exchange.getExchangeSpecification().setExchangeSpecificParametersItem(TICKER_STRATEGY_KEY, tickerStrategies.get("parallelTickerStrategy"));
-                } catch (IOException e) {
-                    LOGGER.debug("IOException fetching tickers for {}: ", exchange.getExchangeSpecification().getExchangeName(), e);
+                if (spread != null) {
+                    trade(spread);
                 }
-            }
-
-            BigDecimal tradingFee = getExchangeFee(exchange, exchangeService.convertExchangePair(exchange, CurrencyPair.BTC_USD), false);
-
-            LOGGER.info("{} ticker strategy: {}",
-                exchange.getExchangeSpecification().getExchangeName(),
-                exchange.getExchangeSpecification().getExchangeSpecificParametersItem(TICKER_STRATEGY_KEY));
-
-            LOGGER.info("{} {} trading fee: {}",
-                exchange.getExchangeSpecification().getExchangeName(),
-                exchangeService.convertExchangePair(exchange, CurrencyPair.BTC_USD),
-                tradingFee);
-        });
-
-        tickerService.initializeTickers(exchanges);
-
-        if (tradingConfiguration.getFixedExposure() != null) {
-            LOGGER.info("Using fixed exposure of ${} as configured", tradingConfiguration.getFixedExposure());
-        }
-
-        if (tradingConfiguration.getTradeTimeout() != null) {
-            LOGGER.info("Using trade timeout of {} hours", tradingConfiguration.getTradeTimeout());
-        }
-
-        // load active trades from file, if there is one
-        File stateFile = new File(STATE_FILE);
-
-        if (stateFile.exists()) {
-            if (!stateFile.canRead()) {
-                LOGGER.error("Cannot read state file: {}", stateFile.getAbsolutePath());
-            } else {
-                try {
-                    activePosition = objectMapper.readValue(stateFile, ActivePosition.class);
-
-                    LOGGER.info("Loaded active trades from file: {}", stateFile.getAbsolutePath());
-                    LOGGER.info("Active trades: {}", activePosition);
-                } catch (IOException e) {
-                    LOGGER.error("Unable to parse state file {}: ", stateFile.getAbsolutePath(), e);
-                }
-            }
-        }
+            });
     }
 
-    /**
-     * As often as once per minute, display a summary of any non-critical error messages. Summarizing them greatly
-     * reduces how noisy the logs are while still providing the same information.
-     */
-    @Scheduled(cron = "0 * * * * *")
-    public void errorSummary() {
-        if (!errorCollectorService.isEmpty()) {
-            errorCollectorService.report().forEach(LOGGER::info);
-            errorCollectorService.clear();
-        }
-    }
+    public void trade(Spread spread) {
+        final String shortExchangeName = spread.getShortExchange().getExchangeSpecification().getExchangeName();
+        final String longExchangeName = spread.getLongExchange().getExchangeSpecification().getExchangeName();
 
-    /**
-     * Display a summary once every 6 hours with the current spreads.
-     */
-    @Scheduled(cron = "0 0 0/6 * * *") // every 6 hours
-    public void summary() {
-        LOGGER.info("Summary: [Long/Short Exchanges] [Pair] [Current Spread] -> [{} Spread Target]", (activePosition != null ? "Exit" : "Entry"));
+        LOGGER.debug("Long/Short: {}/{} {} {}",
+            longExchangeName,
+            shortExchangeName,
+            spread.getCurrencyPair(),
+            spread.getIn());
 
-        List<TradeCombination> tradeCombinations = tickerService.getTradeCombinations();
-
-        tradeCombinations.forEach(tradeCombination -> {
-            Spread spread = computeSpread(tradeCombination);
-
-            if (spread == null) {
+        if (!conditionService.isForceCloseCondition() && spread.getIn().compareTo(tradingConfiguration.getEntrySpread()) > 0) {
+            if (activePosition != null) {
                 return;
             }
 
-            if (activePosition == null && BigDecimal.ZERO.compareTo(spread.getIn()) < 0) {
-                LOGGER.info("{}/{} {} {} -> {}",
-                    spread.getLongExchange().getExchangeSpecification().getExchangeName(),
-                    spread.getShortExchange().getExchangeSpecification().getExchangeName(),
-                    spread.getCurrencyPair(),
-                    spread.getIn(),
-                    tradingConfiguration.getEntrySpread());
-            } else if (activePosition != null
-                && activePosition.getCurrencyPair().equals(spread.getCurrencyPair())
-                && activePosition.getLongTrade().getExchange().equals(spread.getLongExchange().getExchangeSpecification().getExchangeName())
-                && activePosition.getShortTrade().getExchange().equals(spread.getShortExchange().getExchangeSpecification().getExchangeName())) {
+            entryPosition(spread, shortExchangeName, longExchangeName);
+        } else if (activePosition != null
+            && spread.getCurrencyPair().equals(activePosition.getCurrencyPair())
+            && longExchangeName.equals(activePosition.getLongTrade().getExchange())
+            && shortExchangeName.equals(activePosition.getShortTrade().getExchange())
+            && (spread.getOut().compareTo(activePosition.getExitTarget()) < 0 || conditionService.isForceCloseCondition() || isTradeExpired())) {
 
-                LOGGER.info("{}/{} {} {} -> {}",
-                    spread.getLongExchange().getExchangeSpecification().getExchangeName(),
-                    spread.getShortExchange().getExchangeSpecification().getExchangeName(),
-                    spread.getCurrencyPair(),
-                    spread.getOut(),
-                    activePosition.getExitTarget());
-            }
-        });
-    }
-
-    @Scheduled(initialDelay = 5000, fixedRate = 3000)
-    public void tick() {
-        LOGGER.debug("Tick");
-
-        if (bailOut) {
-            LOGGER.error("Exiting immediately to avoid erroneous trades.");
-            System.exit(1);
-        }
-
-        if (activePosition == null && conditionService.isExitWhenIdleCondition()) {
-            LOGGER.info("Exiting at user request");
-            conditionService.clearExitWhenIdleCondition();
-            System.exit(0);
-        }
-
-        tickerService.refreshTickers();
-
-        long exchangePollStartTime = System.currentTimeMillis();
-        List<TradeCombination> tradeCombinations = tickerService.getTradeCombinations();
-
-        tradeCombinations.forEach(tradeCombination -> {
-            Spread spread = computeSpread(tradeCombination);
-
-            if (spread == null) {
-                return;
-            }
-
-            final String shortExchangeName = spread.getShortExchange().getExchangeSpecification().getExchangeName();
-            final String longExchangeName = spread.getLongExchange().getExchangeSpecification().getExchangeName();
-
-            LOGGER.debug("Long/Short: {}/{} {} {}",
-                longExchangeName,
-                shortExchangeName,
-                spread.getCurrencyPair(),
-                spread.getIn());
-
-            if (!conditionService.isForceCloseCondition() && spread.getIn().compareTo(tradingConfiguration.getEntrySpread()) > 0) {
-                if (activePosition != null) {
-                    return;
-                }
-
-                entryPosition(spread, shortExchangeName, longExchangeName);
-            } else if (activePosition != null
-                    && spread.getCurrencyPair().equals(activePosition.getCurrencyPair())
-                    && longExchangeName.equals(activePosition.getLongTrade().getExchange())
-                    && shortExchangeName.equals(activePosition.getShortTrade().getExchange())
-                    && (spread.getOut().compareTo(activePosition.getExitTarget()) < 0 || conditionService.isForceCloseCondition() || isTradeExpired())) {
-
-                exitPosition(spread, shortExchangeName, longExchangeName);
-            }
-        });
-
-        long exchangePollDuration = System.currentTimeMillis() - exchangePollStartTime;
-
-        if (exchangePollDuration > 3000) {
-            LOGGER.warn("Polling exchanges took {} ms", exchangePollDuration);
+            exitPosition(spread, shortExchangeName, longExchangeName);
         }
     }
 
     private void entryPosition(Spread spread, String shortExchangeName, String longExchangeName) {
-        final BigDecimal longFees = getExchangeFee(spread.getLongExchange(), spread.getCurrencyPair(), true);
-        final BigDecimal shortFees = getExchangeFee(spread.getShortExchange(), spread.getCurrencyPair(), true);
+        final BigDecimal longFees = exchangeService.getExchangeFee(spread.getLongExchange(), spread.getCurrencyPair(), true);
+        final BigDecimal shortFees = exchangeService.getExchangeFee(spread.getShortExchange(), spread.getCurrencyPair(), true);
         final CurrencyPair currencyPairLongExchange = exchangeService.convertExchangePair(spread.getLongExchange(), spread.getCurrencyPair());
         final CurrencyPair currencyPairShortExchange = exchangeService.convertExchangePair(spread.getShortExchange(), spread.getCurrencyPair());
 
@@ -407,7 +184,7 @@ public class TradingService {
             return;
         }
 
-        BigDecimal spreadVerification = computeSpread(longLimitPrice, shortLimitPrice);
+        BigDecimal spreadVerification = spreadService.computeSpread(longLimitPrice, shortLimitPrice);
 
         if (spreadVerification.compareTo(tradingConfiguration.getEntrySpread()) < 0) {
             LOGGER.debug("Not enough liquidity to execute both trades profitably");
@@ -423,33 +200,33 @@ public class TradingService {
 
         BigDecimal totalBalance = logCurrentExchangeBalances(spread.getLongExchange(), spread.getShortExchange());
 
-        try {
-            activePosition = new ActivePosition();
-            activePosition.setEntryTime(OffsetDateTime.now());
-            activePosition.setCurrencyPair(spread.getCurrencyPair());
-            activePosition.setExitTarget(exitTarget);
-            activePosition.setEntryBalance(totalBalance);
-            activePosition.getLongTrade().setExchange(spread.getLongExchange());
-            activePosition.getLongTrade().setVolume(longVolume);
-            activePosition.getLongTrade().setEntry(longLimitPrice);
-            activePosition.getShortTrade().setExchange(spread.getShortExchange());
-            activePosition.getShortTrade().setVolume(shortVolume);
-            activePosition.getShortTrade().setEntry(shortLimitPrice);
+//        try {
+        activePosition = new ActivePosition();
+        activePosition.setEntryTime(OffsetDateTime.now());
+        activePosition.setCurrencyPair(spread.getCurrencyPair());
+        activePosition.setExitTarget(exitTarget);
+        activePosition.setEntryBalance(totalBalance);
+        activePosition.getLongTrade().setExchange(spread.getLongExchange());
+        activePosition.getLongTrade().setVolume(longVolume);
+        activePosition.getLongTrade().setEntry(longLimitPrice);
+        activePosition.getShortTrade().setExchange(spread.getShortExchange());
+        activePosition.getShortTrade().setVolume(shortVolume);
+        activePosition.getShortTrade().setEntry(shortLimitPrice);
 
-            executeOrderPair(
-                    spread.getLongExchange(), spread.getShortExchange(),
-                    spread.getCurrencyPair(),
-                    longLimitPrice, shortLimitPrice,
-                    longVolume, shortVolume,
-                    true);
+//            executeOrderPair(
+//                    spread.getLongExchange(), spread.getShortExchange(),
+//                    spread.getCurrencyPair(),
+//                    longLimitPrice, shortLimitPrice,
+//                    longVolume, shortVolume,
+//                    true);
 
-            notificationService.sendEmailNotificationBodyForEntryTrade(spread, exitTarget, longVolume,
-                    longLimitPrice, shortVolume, shortLimitPrice);
+        notificationService.sendEmailNotificationBodyForEntryTrade(spread, exitTarget, longVolume,
+            longLimitPrice, shortVolume, shortLimitPrice);
 
-            } catch (IOException e) {
-                LOGGER.error("IOE executing limit orders: ", e);
-                activePosition = null;
-            }
+//            } catch (IOException e) {
+//                LOGGER.error("IOE executing limit orders: ", e);
+//                activePosition = null;
+//            }
 
         try {
             FileUtils.write(new File(STATE_FILE), objectMapper.writeValueAsString(activePosition), Charset.defaultCharset());
@@ -524,7 +301,7 @@ public class TradingService {
 
         LOGGER.debug("Limit prices: {}/{}", longLimitPrice, shortLimitPrice);
 
-        BigDecimal spreadVerification = computeSpread(longLimitPrice, shortLimitPrice);
+        BigDecimal spreadVerification = spreadService.computeSpread(longLimitPrice, shortLimitPrice);
 
         LOGGER.debug("Spread verification: {}", spreadVerification);
 
@@ -554,36 +331,36 @@ public class TradingService {
 
         logExitTrade();
 
-        try {
-            LOGGER.info("Long close: {} {} {} @ {} ({} slip) = {}{}",
-                longExchangeName,
-                spread.getCurrencyPair(),
-                longVolume,
-                longLimitPrice,
-                longLimitPrice.subtract(spread.getLongTicker().getBid()),
-                Currency.USD.getSymbol(),
-                longVolume.multiply(spread.getLongTicker().getBid()));
-            LOGGER.info("Short close: {} {} {} @ {} ({} slip) = {}{}",
-                shortExchangeName,
-                spread.getCurrencyPair(),
-                shortVolume,
-                shortLimitPrice,
-                spread.getShortTicker().getAsk().subtract(shortLimitPrice),
-                Currency.USD.getSymbol(),
-                shortVolume.multiply(spread.getShortTicker().getAsk()));
+//        try {
+        LOGGER.info("Long close: {} {} {} @ {} ({} slip) = {}{}",
+            longExchangeName,
+            spread.getCurrencyPair(),
+            longVolume,
+            longLimitPrice,
+            longLimitPrice.subtract(spread.getLongTicker().getBid()),
+            Currency.USD.getSymbol(),
+            longVolume.multiply(spread.getLongTicker().getBid()));
+        LOGGER.info("Short close: {} {} {} @ {} ({} slip) = {}{}",
+            shortExchangeName,
+            spread.getCurrencyPair(),
+            shortVolume,
+            shortLimitPrice,
+            spread.getShortTicker().getAsk().subtract(shortLimitPrice),
+            Currency.USD.getSymbol(),
+            shortVolume.multiply(spread.getShortTicker().getAsk()));
 
-            executeOrderPair(
-                spread.getLongExchange(), spread.getShortExchange(),
-                spread.getCurrencyPair(),
-                longLimitPrice, shortLimitPrice,
-                longVolume, shortVolume,
-                false);
-        } catch (IOException e) {
-            LOGGER.error("IOE executing limit orders: ", e);
-            // TODO: Why don't we return here? Could it be because the IOException could be, for example, a timeout exception
-            // and we don't know for sure whether the trade was accepted or not?
-            // Maybe we should return here and send a notification so the user is aware of this situation
-        }
+//            executeOrderPair(
+//                spread.getLongExchange(), spread.getShortExchange(),
+//                spread.getCurrencyPair(),
+//                longLimitPrice, shortLimitPrice,
+//                longVolume, shortVolume,
+//                false);
+//        } catch (IOException e) {
+//            LOGGER.error("IOE executing limit orders: ", e);
+//            // TODO: Why don't we return here? Could it be because the IOException could be, for example, a timeout exception
+//            // and we don't know for sure whether the trade was accepted or not?
+//            // Maybe we should return here and send a notification so the user is aware of this situation
+//        }
 
         LOGGER.info("Combined account balances on entry: ${}", activePosition.getEntryBalance());
         BigDecimal updatedBalance = logCurrentExchangeBalances(spread.getLongExchange(), spread.getShortExchange());
@@ -681,77 +458,6 @@ public class TradingService {
         return minimumAmount;
     }
 
-    private BigDecimal getExchangeFee(Exchange exchange, CurrencyPair currencyPair, boolean isQuiet) {
-        BigDecimal cachedFee = feeCache.getCachedFee(exchange, currencyPair);
-
-        if (cachedFee != null) {
-            return cachedFee;
-        }
-
-        // if an explicit override is configured, default to that
-        if (exchangeService.getExchangeMetadata(exchange).getFeeOverride() != null) {
-            BigDecimal fee = exchangeService.getExchangeMetadata(exchange).getFeeOverride();
-            feeCache.setCachedFee(exchange, currencyPair, fee);
-
-            LOGGER.trace("Using explicitly configured fee override of {} for {}",
-                fee,
-                exchange.getExchangeSpecification().getExchangeName());
-
-            return fee;
-        }
-
-        try {
-            Map<CurrencyPair, Fee> fees = exchange.getAccountService().getDynamicTradingFees();
-
-            if (fees.containsKey(currencyPair)) {
-                BigDecimal fee = fees.get(currencyPair).getMakerFee();
-
-                // We're going to cache this value. Fees don't change all that often and we don't want to use up
-                // our allowance of API calls just checking the fees.
-                feeCache.setCachedFee(exchange, currencyPair, fee);
-
-                LOGGER.trace("Using dynamic maker fee for {}",
-                    exchange.getExchangeSpecification().getExchangeName());
-
-                return fee;
-            }
-        } catch (NotYetImplementedForExchangeException e) {
-            LOGGER.trace("Dynamic fees not yet implemented for {}, will try other methods",
-                    exchange.getExchangeSpecification().getExchangeName());
-        } catch (IOException e) {
-            LOGGER.trace("IOE fetching dynamic trading fees for {}",
-                    exchange.getExchangeSpecification().getExchangeName());
-        } catch (Exception e) {
-            LOGGER.warn("Programming error in XChange! {} when calling getDynamicTradingFees() for exchange: {}",
-                    e.getClass().getName(),
-                    exchange.getExchangeSpecification().getExchangeName());
-        }
-
-        CurrencyPairMetaData currencyPairMetaData = exchange.getExchangeMetaData().getCurrencyPairs().get(exchangeService.convertExchangePair(exchange, currencyPair));
-
-        if (currencyPairMetaData == null || currencyPairMetaData.getTradingFee() == null) {
-            BigDecimal configuredFee = exchangeService.getExchangeMetadata(exchange).getFee();
-
-            if (configuredFee == null) {
-                if (!isQuiet) {
-                    LOGGER.error("{} has no fees configured. Setting default of 0.0030. Please configure the correct value!",
-                            exchange.getExchangeSpecification().getExchangeName());
-                }
-
-                return new BigDecimal("0.0030");
-            }
-
-            if (!isQuiet) {
-                LOGGER.warn("{} fees unavailable via API. Will use configured value.",
-                        exchange.getExchangeSpecification().getExchangeName());
-            }
-
-            return configuredFee;
-        }
-
-        return currencyPairMetaData.getTradingFee();
-    }
-
     private void executeOrderPair(Exchange longExchange, Exchange shortExchange,
                                   CurrencyPair currencyPair,
                                   BigDecimal longLimitPrice, BigDecimal shortLimitPrice,
@@ -759,22 +465,22 @@ public class TradingService {
                                   boolean isPositionOpen) throws IOException, ExchangeException {
 
         LimitOrder longLimitOrder = new LimitOrder.Builder(isPositionOpen ? Order.OrderType.BID : Order.OrderType.ASK, exchangeService.convertExchangePair(longExchange, currencyPair))
-                .limitPrice(longLimitPrice)
-                .originalAmount(longVolume)
-                .build();
+            .limitPrice(longLimitPrice)
+            .originalAmount(longVolume)
+            .build();
         LimitOrder shortLimitOrder = new LimitOrder.Builder(isPositionOpen ? Order.OrderType.ASK : Order.OrderType.BID, exchangeService.convertExchangePair(shortExchange, currencyPair))
-                .limitPrice(shortLimitPrice)
-                .originalAmount(shortVolume)
-                .build();
+            .limitPrice(shortLimitPrice)
+            .originalAmount(shortVolume)
+            .build();
 
         shortLimitOrder.setLeverage("2");
 
         LOGGER.debug("{}: {}",
-                longExchange.getExchangeSpecification().getExchangeName(),
-                longLimitOrder);
+            longExchange.getExchangeSpecification().getExchangeName(),
+            longLimitOrder);
         LOGGER.debug("{}: {}",
-                shortExchange.getExchangeSpecification().getExchangeName(),
-                shortLimitOrder);
+            shortExchange.getExchangeSpecification().getExchangeName(),
+            shortLimitOrder);
 
         try {
             String longOrderId = longExchange.getTradeService().placeLimitOrder(longLimitOrder);
@@ -856,13 +562,6 @@ public class TradingService {
         return Optional.empty();
     }
 
-    private BigDecimal computeSpread(BigDecimal longPrice, BigDecimal shortPrice) {
-        BigDecimal scaledLongPrice = longPrice.setScale(BTC_SCALE, RoundingMode.HALF_EVEN);
-        BigDecimal scaledShortPrice = shortPrice.setScale(BTC_SCALE, RoundingMode.HALF_EVEN);
-
-        return (scaledShortPrice.subtract(scaledLongPrice)).divide(scaledLongPrice, RoundingMode.HALF_EVEN);
-    }
-
     BigDecimal getVolumeForOrder(Exchange exchange, CurrencyPair currencyPair, String orderId, BigDecimal defaultVolume) {
         try {
             LOGGER.debug("{}: Attempting to fetch volume from order by ID: {}", exchange.getExchangeSpecification().getExchangeName(), orderId);
@@ -896,7 +595,7 @@ public class TradingService {
         }
 
         try {
-            BigDecimal balance = getAccountBalance(exchange, currencyPair.base);
+            BigDecimal balance = exchangeService.getAccountBalance(exchange, currencyPair.base);
 
             if (BigDecimal.ZERO.compareTo(balance) < 0) {
                 LOGGER.debug("{}: Using {} balance: {}", exchange.getExchangeSpecification().getExchangeName(), currencyPair.base.toString(), balance);
@@ -946,10 +645,10 @@ public class TradingService {
                 .parallel()
                 .map(exchange -> {
                     try {
-                        return getAccountBalance(exchange);
+                        return exchangeService.getAccountBalance(exchange);
                     } catch (IOException e) {
                         LOGGER.trace("IOException fetching {} account balance",
-                                exchange.getExchangeSpecification().getExchangeName());
+                            exchange.getExchangeSpecification().getExchangeName());
                     }
 
                     return BigDecimal.ZERO;
@@ -958,8 +657,8 @@ public class TradingService {
                 .orElse(BigDecimal.ZERO);
 
             BigDecimal exposure = smallestBalance
-                    .multiply(TRADE_PORTION)
-                    .setScale(DecimalConstants.USD_SCALE, RoundingMode.HALF_EVEN);
+                .multiply(TRADE_PORTION)
+                .setScale(DecimalConstants.USD_SCALE, RoundingMode.HALF_EVEN);
 
             LOGGER.debug("Maximum exposure for {}: {}", exchanges, exposure);
 
@@ -969,16 +668,16 @@ public class TradingService {
 
     private BigDecimal logCurrentExchangeBalances(Exchange longExchange, Exchange shortExchange) {
         try {
-            BigDecimal longBalance = getAccountBalance(longExchange);
-            BigDecimal shortBalance = getAccountBalance(shortExchange);
+            BigDecimal longBalance = exchangeService.getAccountBalance(longExchange);
+            BigDecimal shortBalance = exchangeService.getAccountBalance(shortExchange);
             BigDecimal totalBalance = longBalance.add(shortBalance);
 
             LOGGER.info("Updated account balances: {} ${} + {} ${} = ${}",
-                    longExchange.getExchangeSpecification().getExchangeName(),
-                    longBalance,
-                    shortExchange.getExchangeSpecification().getExchangeName(),
-                    shortBalance,
-                    totalBalance);
+                longExchange.getExchangeSpecification().getExchangeName(),
+                longBalance,
+                shortExchange.getExchangeSpecification().getExchangeName(),
+                shortBalance,
+                totalBalance);
 
             return totalBalance;
         } catch (IOException e) {
@@ -986,33 +685,6 @@ public class TradingService {
         }
 
         return BigDecimal.ZERO;
-    }
-
-    BigDecimal getAccountBalance(Exchange exchange, Currency currency, int scale) throws IOException {
-        AccountService accountService = exchange.getAccountService();
-
-        for (Wallet wallet : accountService.getAccountInfo().getWallets().values()) {
-            if (wallet.getBalances().containsKey(currency)) {
-                return wallet.getBalance(currency).getAvailable()
-                        .setScale(scale, RoundingMode.HALF_EVEN);
-            }
-        }
-
-        LOGGER.error("{}: Unable to fetch {} balance",
-            exchange.getExchangeSpecification().getExchangeName(),
-            currency.getCurrencyCode());
-
-        return BigDecimal.ZERO;
-    }
-
-    private BigDecimal getAccountBalance(Exchange exchange, Currency currency) throws IOException {
-        return getAccountBalance(exchange, currency, USD_SCALE);
-    }
-
-    private BigDecimal getAccountBalance(Exchange exchange) throws IOException {
-        Currency currency = exchangeService.getExchangeHomeCurrency(exchange);
-
-        return getAccountBalance(exchange, currency);
     }
 
     /*
@@ -1042,34 +714,6 @@ public class TradingService {
         return activePosition.getEntryTime().plusHours(tradingConfiguration.getTradeTimeout()).isBefore(OffsetDateTime.now());
     }
 
-    private Spread computeSpread(TradeCombination tradeCombination) {
-        Exchange longExchange = tradeCombination.getLongExchange();
-        Exchange shortExchange = tradeCombination.getShortExchange();
-        CurrencyPair currencyPair = tradeCombination.getCurrencyPair();
-
-        Ticker longTicker = tickerService.getTicker(longExchange, currencyPair);
-        Ticker shortTicker = tickerService.getTicker(shortExchange, currencyPair);
-
-        if (tickerService.isInvalidTicker(longTicker) || tickerService.isInvalidTicker(shortTicker)) {
-            return null;
-        }
-
-        BigDecimal spreadIn = computeSpread(longTicker.getAsk(), shortTicker.getBid());
-        BigDecimal spreadOut = computeSpread(longTicker.getBid(), shortTicker.getAsk());
-
-        Spread spread = new Spread(
-            currencyPair,
-            longExchange,
-            shortExchange,
-            longTicker,
-            shortTicker,
-            spreadIn,
-            spreadOut);
-
-        spreadService.publish(spread);
-
-        return spread;
-    }
 
     protected void persistArbitrageToCsvFile(ArbitrageLog arbitrageLog) {
         final File csvFile = new File(TRADE_HISTORY_FILE);
