+++ src/main/java/com/r307/arbitrader/service/ticker/StreamingTickerStrategy.java
@@ -2,6 +2,7 @@ package com.r307.arbitrader.service.ticker;
 
 import com.r307.arbitrader.service.ErrorCollectorService;
 import com.r307.arbitrader.service.ExchangeService;
+import com.r307.arbitrader.service.TradingScheduler;
 import info.bitrich.xchangestream.core.ProductSubscription;
 import info.bitrich.xchangestream.core.StreamingExchange;
 import info.bitrich.xchangestream.gemini.GeminiStreamingExchange;
@@ -28,11 +29,13 @@ public class StreamingTickerStrategy implements TickerStrategy {
     private final Map<StreamingExchange, Map<CurrencyPair, Ticker>> tickers = new HashMap<>();
     private final ErrorCollectorService errorCollectorService;
     private final ExchangeService exchangeService;
+    private final TradingScheduler tradingScheduler;
 
     @Inject
-    public StreamingTickerStrategy(ErrorCollectorService errorCollectorService, ExchangeService exchangeService) {
+    public StreamingTickerStrategy(ErrorCollectorService errorCollectorService, ExchangeService exchangeService, TradingScheduler tradingScheduler) {
         this.errorCollectorService = errorCollectorService;
         this.exchangeService = exchangeService;
+        this.tradingScheduler = tradingScheduler;
     }
 
     @Override
@@ -71,28 +74,35 @@ public class StreamingTickerStrategy implements TickerStrategy {
                 // TODO: Temp fix while https://github.com/knowm/XChange/pull/3761 is not merged and released
                 final Observable<Ticker> tickerObservable;
                 if (exchange instanceof GeminiStreamingExchange) {
-                    tickerObservable = exchange.getStreamingMarketDataService().getTicker(exchangeService.convertExchangePair(exchange, pair), 10);
+                    tickerObservable = exchange.getStreamingMarketDataService().getTicker(exchangeService.convertExchangePair(exchange, pair), 100);
                 }
                 else {
                     tickerObservable = exchange.getStreamingMarketDataService().getTicker(exchangeService.convertExchangePair(exchange, pair));
                 }
-                return tickerObservable.subscribe(
-                    ticker -> {
-                        tickers.computeIfAbsent(exchange, e -> new HashMap<>());
-                        tickers.get(exchange).put(pair, ticker);
-
-                        LOGGER.debug("Received ticker: {} {} {}/{}",
-                            exchange.getExchangeSpecification().getExchangeName(),
-                            ticker.getCurrencyPair(),
-                            ticker.getBid(),
-                            ticker.getAsk());
-                    },
-                    throwable -> {
-                        errorCollectorService.collect(exchange, throwable);
-                        LOGGER.debug("Unexpected checked exception: {}", throwable.getMessage(), throwable);
+                return tickerObservable
+                    .map(ticker -> getTicker(exchange, pair, ticker))
+                    .doOnNext(ticker -> onNext(exchange, ticker))
+                    .subscribe(
+                        ticker -> tradingScheduler.startTradingProcess(),
+                        throwable -> {
+                            errorCollectorService.collect(exchange, throwable);
+                            LOGGER.debug("Unexpected checked exception: {}", throwable.getMessage(), throwable);
                     });
             })
             .collect(Collectors.toList());
+    }
+
+    private Ticker getTicker(StreamingExchange exchange, CurrencyPair pair, Ticker ticker) {
+        tickers.computeIfAbsent(exchange, e -> new HashMap<>());
+        return tickers.get(exchange).put(pair, ticker);
+    }
+
+    private void onNext(StreamingExchange exchange, Ticker ticker) {
+        LOGGER.debug("Received ticker: {} {} {}/{}",
+            exchange.getExchangeSpecification().getExchangeName(),
+            ticker.getCurrencyPair(),
+            ticker.getBid(),
+            ticker.getAsk());
     }
 
     @Override
