+++ src/test/java/com/r307/arbitrader/service/ticker/ParallelTickerStrategyTest.java
@@ -3,18 +3,22 @@ package com.r307.arbitrader.service.ticker;
 import com.r307.arbitrader.ExchangeBuilder;
 import com.r307.arbitrader.config.NotificationConfiguration;
 import com.r307.arbitrader.service.ErrorCollectorService;
+import com.r307.arbitrader.service.ExchangeFeeCache;
 import com.r307.arbitrader.service.ExchangeService;
+import com.r307.arbitrader.service.TickerService;
 import org.junit.Before;
 import org.junit.Test;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.marketdata.Ticker;
 import org.knowm.xchange.exceptions.ExchangeException;
+import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
 import java.io.IOException;
 import java.lang.reflect.UndeclaredThrowableException;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
 
 import static org.junit.Assert.*;
@@ -31,7 +35,7 @@ public class ParallelTickerStrategyTest {
         MockitoAnnotations.initMocks(this);
 
         NotificationConfiguration notificationConfiguration = new NotificationConfiguration();
-        ExchangeService exchangeService = new ExchangeService();
+        ExchangeService exchangeService = new ExchangeService(new HashMap<>(), new ExchangeFeeCache());
 
         errorCollectorService = new ErrorCollectorService();
