+++ src/test/java/com/r307/arbitrader/service/ticker/SingleCallTickerStrategyTest.java
@@ -10,6 +10,7 @@ import org.knowm.xchange.Exchange;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.marketdata.Ticker;
 import org.knowm.xchange.exceptions.ExchangeException;
+import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import java.io.IOException;
@@ -26,12 +27,14 @@ public class SingleCallTickerStrategyTest {
 
     private TickerStrategy tickerStrategy;
 
+    @Mock
+    private ExchangeService exchangeService;
+
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
 
         NotificationConfiguration notificationConfiguration = new NotificationConfiguration();
-        ExchangeService exchangeService = new ExchangeService();
 
         errorCollectorService = new ErrorCollectorService();
