+++ src/test/java/com/r307/arbitrader/service/ticker/StreamingTickerStrategyTest.java
@@ -3,12 +3,14 @@ package com.r307.arbitrader.service.ticker;
 import com.r307.arbitrader.ExchangeBuilder;
 import com.r307.arbitrader.service.ErrorCollectorService;
 import com.r307.arbitrader.service.ExchangeService;
+import com.r307.arbitrader.service.TradingScheduler;
 import org.junit.Before;
 import org.junit.Test;
 import org.knowm.xchange.Exchange;
 import org.knowm.xchange.currency.CurrencyPair;
 import org.knowm.xchange.dto.marketdata.Ticker;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 
 import java.util.Collections;
@@ -27,8 +29,9 @@ public class StreamingTickerStrategyTest {
     @Before
     public void setUp() {
         MockitoAnnotations.initMocks(this);
+        final TradingScheduler tradingSchedulerMock = Mockito.mock(TradingScheduler.class);
 
-        streamingTickerStrategy = new StreamingTickerStrategy(errorCollectorService, exchangeService);
+        streamingTickerStrategy = new StreamingTickerStrategy(errorCollectorService, exchangeService, tradingSchedulerMock);
     }
 
     @Test
